CAPI-217(chore): use ComponentManager to create VideoManager class
