+++ src/screens/videoStream/index.ts
@@ -6,41 +6,54 @@ import createOverlayElement from './overlay';
 import css from './index.module.css';
 import StreamManager from '../../modules/StreamManager';
 import createSpinner from './spinner';
+import ComponentManager from '../../modules/ComponentManager';
+import { getNonWideAngleCamera } from '../../camera/getNonWideAngleCamera';
+import VideoManager from '../../modules/VideoManager';
 
-export default function createContainerElement(): HTMLDivElement {
-	const container = document.createElement('div');
-	container.className = css.container;
+export default class VideoStreamScreen extends ComponentManager {
+	constructor() {
+		super();
+		const container = super.getElement();
 
-	const spinner = createSpinner();
-	container.appendChild(spinner);
+		container.className = css.container;
 
-	const streamManager = StreamManager.getInstance();
+		// initially show loader
+		const spinner = createSpinner();
+		container.appendChild(spinner);
 
-	streamManager.onStreamSet(() => {
-		container.removeChild(spinner);
+		super.onMount(async () => {
+			const device = await getNonWideAngleCamera();
+			const streamManager = StreamManager.getInstance();
+			const stream = await streamManager.getStream(device);
+			const videoManager = VideoManager.getInstance();
+			const videoElement = videoManager.getVideoElement();
+			videoElement.srcObject = stream;
 
-		// attach tire overlay
-		const overlayElement = createOverlayElement();
+			// create tire overlay
+			const overlayElement = createOverlayElement();
 
-		// attach close sdk button
-		const closeElement = createCloseElement();
+			// create close sdk button
+			const closeElement = createCloseElement();
 
-		// create instructions
-		const instructionsElement = createInstructionElement();
+			// create instructions
+			const instructionsElement = createInstructionElement();
 
-		// create capture button
-		const captureButton = createButtonElement();
+			// create capture button
+			const captureButton = createButtonElement();
 
-		// attach bottom section
-		const bottomSection = createBottomSection(
-			instructionsElement,
-			captureButton
-		);
+			// create bottom section
+			const bottomSection = createBottomSection(
+				instructionsElement,
+				captureButton
+			);
 
-		container.appendChild(overlayElement);
-		container.appendChild(closeElement);
-		container.appendChild(bottomSection);
-	});
+			// hide loader
+			container.removeChild(spinner);
 
-	return container;
+			container.appendChild(videoElement);
+			container.appendChild(overlayElement);
+			container.appendChild(closeElement);
+			container.appendChild(bottomSection);
+		});
+	}
 }
